avoid using RefCell and refactor futhrer
